requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
python-dotenv>=1.0.0
openai>=1.3.7 
pandas>=2.1.1 
rich>=13.5.2